ML/pyproject.toml
[project]
name = "FakeNewsDetector"
version = "0.1.0"
description = "虚假新闻检测系统"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.2.0",
    "jieba>=0.42.1",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "seaborn>=0.12.0"
]

[project.scripts]
fakenewsdetector = "main:main"

[project.optional-dependencies]
dev = [
    "hatchling>=1.18.0",
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0"
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest]
testpaths = ["tests"]
